integrate streams into decomposed metacom
