+++ lib/channel.js
@@ -1,41 +1,16 @@
 'use strict';
 
 const http = require('http');
-const path = require('path');
 const metautil = require('metautil');
-const { MetacomReadable, MetacomWritable, MetacomChunk } = require('./streams');
-
-const MIME_TYPES = {
-  html: 'text/html; charset=UTF-8',
-  json: 'application/json; charset=UTF-8',
-  js: 'application/javascript; charset=UTF-8',
-  css: 'text/css',
-  png: 'image/png',
-  ico: 'image/x-icon',
-  svg: 'image/svg+xml',
-};
-
-const HEADERS = {
-  'X-XSS-Protection': '1; mode=block',
-  'X-Content-Type-Options': 'nosniff',
-  'Strict-Transport-Security': 'max-age=31536000; includeSubdomains; preload',
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Methods': 'POST, GET, OPTIONS',
-  'Access-Control-Allow-Headers': 'Content-Type',
-};
-
-const TOKEN = 'token';
-const EPOCH = 'Thu, 01 Jan 1970 00:00:00 GMT';
-const FUTURE = 'Fri, 01 Jan 2100 00:00:00 GMT';
-const LOCATION = 'Path=/; Domain';
-const COOKIE_DELETE = `${TOKEN}=deleted; Expires=${EPOCH}; ${LOCATION}=`;
-const COOKIE_HOST = `Expires=${FUTURE}; ${LOCATION}`;
+const transport = require('./transport.js');
+const { MIME_TYPES, HEADERS } = transport.http;
+const { COOKIE_DELETE, COOKIE_HOST, TOKEN } = transport.http;
+const { MetacomReadable, MetacomWritable, MetacomChunk } = require('./streams.js');
 
 class Session {
   constructor(token, channel, data) {
     this.token = token;
     this.channel = channel;
-    this.channels = new Map([channel]);
     this.data = data;
     this.context = new Proxy(data, {
       get: (data, key) => {
@@ -52,8 +27,8 @@ class Session {
   }
 }
 
-const sessions = new Map();
-const channels = new Map();
+const sessions = new Map(); // token: Session
+const channels = new Map(); // Client: Channel
 
 class Client {
   constructor() {
@@ -115,6 +90,7 @@ class Client {
 
   startSession(token, data = {}) {
     const channel = channels.get(this);
+    if (channel.session) sessions.delete(channel.session.token);
     const session = new Session(token, channel, data);
     channel.session = session;
     sessions.set(token, session);
@@ -148,28 +124,18 @@ class Channel {
     return this.session.token;
   }
 
-  static() {
-    const { req, res, ip, application } = this;
-    const { url, method } = req;
-    const filePath = url === '/' ? '/index.html' : url;
-    const fileExt = path.extname(filePath).substring(1);
-    const mimeType = MIME_TYPES[fileExt] || MIME_TYPES.html;
-    res.writeHead(200, { ...HEADERS, 'Content-Type': mimeType });
-    if (res.writableEnded) return;
-    const data = application.getStaticFile(filePath);
-    if (data) {
-      res.end(data);
-      application.console.log(`${ip}\t${method}\t${url}`);
-      return;
-    }
-    this.error(404);
-  }
-
   redirect(location) {
     const { res } = this;
     if (res.headersSent) return;
     res.writeHead(302, { Location: location, ...HEADERS });
     res.end();
+  }
+
+  send(ext, data) {
+    const { res } = this;
+    const mimeType = MIME_TYPES[ext] || MIME_TYPES.html;
+    res.writeHead(200, { ...HEADERS, 'Content-Type': mimeType });
+    res.end(data);
   }
 
   options() {
@@ -365,10 +331,10 @@ class Channel {
   }
 
   destroy() {
+    channels.delete(this.client);
     for (const callback of this.client.events.close) callback();
     if (!this.session) return;
-    const token = this.session.token;
-    sessions.delete(token);
+    sessions.delete(this.session.token);
   }
 }
