+++ lib/server.js
@@ -3,12 +3,11 @@
 const http = require('http');
 const https = require('https');
 const { threadId } = require('worker_threads');
-
 const metautil = require('metautil');
 const { Semaphore } = metautil;
 const ws = require('ws');
-
 const { Channel, channels } = require('./channel.js');
+const { serveStatic } = require('./static.js');
 
 const SHORT_TIMEOUT = 500;
 
@@ -62,7 +61,6 @@ class Server {
         else channel.message(data.toString());
       });
       connection.on('close', () => {
-        channels.delete(channel.client);
         channel.destroy();
       });
     });
@@ -80,12 +78,10 @@ class Server {
     let finished = false;
     const { url } = req;
     const channel = await new Channel(req, res, null, this.application);
-    const { client } = channel;
 
     res.on('close', () => {
       if (finished) return;
       finished = true;
-      channels.delete(client);
       channel.destroy();
     });
 
@@ -98,7 +94,7 @@ class Server {
     }
 
     if (url.startsWith('/api')) this.request(channel);
-    else channel.static();
+    else serveStatic(channel);
   }
 
   request(channel) {
