+++ lib/streams.js
@@ -1,4 +1,5 @@
 'use strict';
+
 const { EventEmitter } = require('events');
 const { Blob } = require('buffer');
