+++ lib/transport.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  http: require('./http.js'),
+};
